nanda_adapter/implementation/requirements.txt
# Core dependencies for langchain_pirate.py
langchain-core
langchain-anthropic
python-dotenv

# NANDA Adapter
nanda-adapter
